fix: rename to devTool/install
