+++ packages/fx-core/resource/yaml.schema.json
@@ -54,7 +54,7 @@
           { "$ref": "#/definitions/botframeworkCreate" },
           { "$ref": "#/definitions/fileCreateOrUpdateEnvironmentFile" },
           { "$ref": "#/definitions/fileCreateOrUpdateJsonFile" },
-          { "$ref": "#/definitions/prerequisiteInstall" },
+          { "$ref": "#/definitions/devToolInstall" },
           { "$ref": "#/definitions/m365TitleAcquire" },
           { "$ref": "#/definitions/spfxDeploy" },
           { "$ref": "#/definitions/teamsAppCopyAppPackageToSPFx" },
@@ -1133,10 +1133,10 @@
         }
       }
     },
-    "prerequisiteInstall": {
+    "devToolInstall": {
       "type": "object",
       "additionalProperties": false,
-      "description": "Install prerequisite(s). Refer to https://aka.ms/teamsfx-actions/prerequisite-install for more details.",
+      "description": "Install development tool(s). Refer to https://aka.ms/teamsfx-actions/devtool-install for more details.",
       "required": ["uses", "with"],
       "properties": {
         "name": {
@@ -1152,8 +1152,8 @@
         },
         "uses": {
           "type": "string",
-          "description": "Install prerequisite(s). Refer to https://aka.ms/teamsfx-actions/prerequisite-install for more details.",
-          "const": "prerequisite/install"
+          "description": "Install development tool(s). Refer to https://aka.ms/teamsfx-actions/devtool-install for more details.",
+          "const": "devTool/install"
         },
         "with": {
           "type": "object",
