+++ packages/fx-core/tests/core/middleware/testAssets/v3Migration/debug/transparent-notification/expected/app.local.yml
@@ -42,10 +42,10 @@ provision:
       appPackagePath: ./build/appPackage/appPackage.${{TEAMSFX_ENV}}.zip # Relative path to this file. This is the path for built zip file.
 
 deploy:
-  - uses: prerequisite/install # Install dependencies
+  - uses: devTool/install # Install development tool(s)
     with:
       func: true
-    writeToEnvironmentFile: # Write the information of installed dependencies into environment file for the specified environment variable(s).
+    writeToEnvironmentFile: # Write the information of installed development tool(s) into environment file for the specified environment variable(s).
       funcPath: FUNC_PATH
 
   - uses: file/createOrUpdateEnvironmentFile # Generate runtime environment variables
