+++ packages/fx-core/tests/core/middleware/testAssets/v3Migration/jsTabWithApi/expectedResult/ts.app.yml
@@ -80,10 +80,10 @@ deploy:
       artifactFolder: build
       # The resource id of the cloud resource to be deployed to. This key will be generated by arm/deploy action automatically. You can replace it with your existing Azure Resource id or add it to your environment variable file.
       resourceId: ${{PROVISIONOUTPUT__AZURESTORAGETABOUTPUT__STORAGERESOURCEID}}
-  - uses: prerequisite/install # Install dependencies
+  - uses: devTool/install # Install development tool(s)
     with:      
       dotnet: true
-    writeToEnvironmentFile: # Write the information of installed dependencies into environment file for the specified environment variable(s).
+    writeToEnvironmentFile: # Write the information of installed development tool(s) into environment file for the specified environment variable(s).
       dotnetPath: DOTNET_PATH
   - uses: cli/runNpmCommand
     name: install api dependencies
@@ -94,7 +94,7 @@ deploy:
     with:
       workingDirectory: api
       args: build extensions.csproj -o bin --ignore-failed-sources
-      execPath: ${{DOTNET_PATH}} # Use dotnet installed by prerequisite/install action
+      execPath: ${{DOTNET_PATH}} # Use dotnet installed by devTool/install action
   - uses: cli/runNpmCommand
     name: build api app
     with:
