+++ templates/scenarios/js/m365-tab/teamsapp.local.yml.tpl
@@ -62,11 +62,11 @@ provision:
       appId: M365_APP_ID
 
 deploy:
-  - uses: prerequisite/install # Install dependencies
+  - uses: devTool/install # Install development tool(s)
     with:
       devCert:
         trust: true
-    writeToEnvironmentFile: # Write the information of installed dependencies into environment file for the specified environment variable(s).
+    writeToEnvironmentFile: # Write the information of installed development tool(s) into environment file for the specified environment variable(s).
       sslCertFile: SSL_CRT_FILE
       sslKeyFile: SSL_KEY_FILE
